// src/app/core/models/usuario.model.ts
import { getDate } from 'ngx-bootstrap/chronos/utils/date-getters';

export class UsuariosI {

    _id: string;
    nombre: String;
    apellido: String;
    sexo: String;
    fotoPerfil: String;
    usuario: String;
    telefono: String;
    email: String;
    fechaN: String;
    password: String;
    estado: Boolean;
    role: String;
}